pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minima"
version = "1.0.0"
description = "Framework modulaire de scraping et d’analyse extensible par plugins."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Unlicense" }
authors = [
    { name = "Isidore La Flamme & ChatGPT" }
]
keywords = ["scraping", "framework", "plugins", "analysis", "python"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Public Domain",
    "Operating System :: OS Independent",
]

dependencies = [
    "beautifulsoup4>=4.14",
    "requests>=2.32",
    "PyYAML>=6.0",
    "tqdm>=4.67",
    "rich>=14.2",
    "typer>=0.20",
    "watchdog>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4",
    "flake8>=7.3",
    "coverage>=7.11"
]

[project.scripts]
minima = "minima.cli:app"

[tool.setuptools]
packages = ["minima"]
include-package-data = true

[tool.setuptools.package-data]
"minima" = ["config/*.yaml", "exports/*", "logs/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.flake8]
max-line-length = 100
ignore = ["E203", "W503"]

[tool.coverage.run]
source = ["minima"]
branch = true
